tidy(ui): organize layer components
