+++ invokeai/frontend/web/src/features/controlLayers/components/CALayer/CALayerOpacity.tsx
