+++ invokeai/frontend/web/src/features/controlLayers/components/IPALayer/IPALayer.tsx
@@ -2,9 +2,9 @@ import { Flex, Spacer, useDisclosure } from '@invoke-ai/ui-library';
 import { createMemoizedSelector } from 'app/store/createMemoizedSelector';
 import { useAppSelector } from 'app/store/storeHooks';
 import ControlAdapterLayerConfig from 'features/controlLayers/components/controlAdapterOverrides/ControlAdapterLayerConfig';
-import { LayerDeleteButton } from 'features/controlLayers/components/LayerDeleteButton';
-import { LayerTitle } from 'features/controlLayers/components/LayerTitle';
-import { LayerVisibilityToggle } from 'features/controlLayers/components/LayerVisibilityToggle';
+import { LayerDeleteButton } from 'features/controlLayers/components/LayerCommon/LayerDeleteButton';
+import { LayerTitle } from 'features/controlLayers/components/LayerCommon/LayerTitle';
+import { LayerVisibilityToggle } from 'features/controlLayers/components/LayerCommon/LayerVisibilityToggle';
 import { isIPAdapterLayer, selectControlLayersSlice } from 'features/controlLayers/store/controlLayersSlice';
 import { memo, useMemo } from 'react';
 import { assert } from 'tsafe';
@@ -13,7 +13,7 @@ type Props = {
   layerId: string;
 };
 
-export const IPLayerListItem = memo(({ layerId }: Props) => {
+export const IPALayer = memo(({ layerId }: Props) => {
   const selector = useMemo(
     () =>
       createMemoizedSelector(selectControlLayersSlice, (controlLayers) => {
@@ -44,4 +44,4 @@ export const IPLayerListItem = memo(({ layerId }: Props) => {
   );
 });
 
-IPLayerListItem.displayName = 'IPLayerListItem';
+IPALayer.displayName = 'IPALayer';
