+++ invokeai/frontend/web/src/features/controlLayers/components/LayerCommon/LayerDeleteButton.tsx
