+++ invokeai/frontend/web/src/features/controlLayers/components/LayerCommon/LayerTitle.tsx
