+++ invokeai/frontend/web/src/features/controlLayers/components/RGLayer/RGLayerColorPicker.tsx
